+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to the plugin.
 ### Added
 - Local *temp variables with the same name as *global variables now generate a warning.
 - *stat_chart commands are now parsed and errors flagged.
+- Nested multireplaces are now flagged as an error.
 
 ### Fixed
 - Experimental array syntax is no longer flagged as an error.
@@ -12,6 +13,7 @@ Changes to the plugin.
 - Definitions now only allowed on actual variable or label references, not any old text in the document.
 - References in *goto and *gosub commands are now indexed.
 - References in *stat_chart are now indexed.
+- Fixed potential endless recursion in parser.
 
 ## [1.0.0] - 2019-12-31
 ### Added